BGE: Restore shadows
When we render shadows, the GPUMaterial is not available, so we can't
update ModelViewMatrix etc... in shadow pass. We'll do it only in main
pass. This was causing a crash when we tried to access material->pass
because material was not NULL but equal to 0xcdcdcdcd.

This can be a temp fix because maybe the ModelViewMatrix has also to be
updated in shadow pass if we don't want to have a gap. But in most cases
this is not noticeable. For now I let it like that.