BGE: Restore part of missing functions for background rendering
World has really weird behaviour for now. There is a big work and a big
refactor planned for world. Blender2.8 devs and maybe aromanov are more concerned by this
area (view3d_draw.c, view3d_draw_legacy, gpu_material.c, shaders for
world, codegen etc....)...

So this should not be our priority for now.

As we just do a quad rendering to draw, we'll have only few things to
change in our code.

I suggest to wait to see the complete refactor so we can adapt our code
 (KX_WorldInfo, API, ImageRender...)