UPBGE: Use KX_GameObject in KX_Scene and object actuators.
Previously both SCA_IObject and CValue were used in KX_Scene and object
actuators like KX_SCA_AddObjectActuator, KX_SCA_EndObjectActuator and
KX_SCA_ReplaceMeshActuator.

The usage of these type needed to use more cast in KX_Scene and was
useless to use in the object actuator which has a fully acces to
KX_GameObject in run and in conversion.