UPBGE: Fix reinstance physics mesh with wire material.
Previously when a mesh using a wire material was converted the
polygons (RAS_Polygons) were added but ths indices were invalid
because we followed the same condition that select to add indices
for polygons or for lines.
This caused to have invalid polygons when reinstancing physics
mesh.

To fix this issue the polygons always get the indices in the same
way for lines of real polygons.

Reported by HG1.