UPBGE: Fix uses lighting flag.
Before refactor of render node and the node arguments the uses
lighting flag was tested at every render of a mesh slot, but now
with render node argument this value was tested only at the
initialization of the arguments.

To solve this issue the flag is tested at the adding of the material
render node in the tree. (GenerateTree function)

Fix issue #497.