UPBGE: Add gccfilter support for compiler messages.
Gccfilter is tool to colorize and format the output of gcc to make them
clearer to human. To use this tool we have to define a compile launch rule
calling gccfilter, this is done only when the option WITH_COLOR_GCC is set
in cmake configuration (disabled as default). The parameter passed to
gccfilter are contained in the variable COLOR_GCC_OPTIONS, for the moment
they are: -c -r -w