UPBGE: Implement node dirty flag for multiple users.
Previously only one flag noticed that the object node trasform changed,
using only one flag is limiting for futur work as occlusion were we would
like to rebuild a BVH tree if a node changed.

To avoid this limitation we use a bitmask where a value of 0 mean not dirty
and a value of 1 dirty. The bitmask is compared through the functions
SG_Node::IsDirty(flag) and cleared thanks to SG_Node::ClearDirty.

The current users are DIRTY_RENDER and DIRTY_CULLING.