fix lighting issues
actually the utilTex was working but in eevee_surface_lit fragment shader
we still have missing uniforms that makes the computation of irradiance
incorrect.

This is why we had dark materials in bge.

Temporarly for this branch I simplify the fragment shader result waiting
we get other uniforms needed for complete irradiance computation.

Now we have something that looks more to eevee (I added only fresnel to
radiance but it looks nice :)

This is a temporarly solution because I added a define in shader defines
that is always here but I don't notice significant changes in viewport
with my basics materials.