UPBGE: Fix depth transprency meshes not sorted.
Previously the depth transparency meshes were rendered after the other
transparency meshes, this break the sorting of transparency and depth
transparency meshes because they are split in two render steps.

To fix this the depth transprency are add in both ALPHA_[...]_BUCKET and
ALPHE_DEPTH_[...]_BUCKET lists and only the ALPHA_[...]_BUCKET list is
rendered which factorize the ordering call, tree generation etc...

Bug reported by Hossein Amiran in e-mail.