add a 2d filter tonemapping
I took the simplest algo I found:
https://learnopengl.com/#!Advanced-Lighting/HDR

The thing is that in eevee the tonemapping is applied in any case (I
think) with another algo as a post processing effect.

Disadvantages of a 2D Filter:

in eevee I guess background color is not treated by default but this
is the tonemap effect that gives the final color to the background I
guess.
In bge, default background color is linearrgb_to_srgb_v4(srgbcolor,
m_horizoncolor.getValue()); and with this, we have exactly the same
color as in eevee. BUT if we apply my filter the background color
changes so maybe a 2D filter is not a good approach.
So we don't have the same colors for background and materials than eevee
(not same image in viewport and in bge)

Advantages: ToneMapping is optionnal. The 2D filter requires to enable HDR RGBA16F
option to work. So even if the algo is simple this costs some
performances so maybe we don't want to have it enabled by default.
With a 2D filter we have flexibility. For example I added a uniform for
exposure with a default value. The user can change it if he wants to
adding a float propoerty on the object which has the 2D Filter.
We can also want to not have tonemapping on our image.

So I don't really know what solution we can have for this issue...

For now I just copied the code and set a random value for exposure.

Well this filter is not very convincing, I'll try other codes or change
values