UPBGE: Reorder libraries linkage.
For the linker the order of the library to link is important because
the first library should only depends on the next library.

BFBGE and UPBGE encountered a lot of issue because of this order and
the mess in macro.cmake doesn't helped to have a clean order.

To clean the order the independent libraries of blender as intern/ or
extern/ are move at the end of the library list. The libraries from
source/blender/ are moved at the center of the list and the game libraries
are moved on top of the list same as editor libraries.