UPBGE: Fix shader used for all transparency material in shadow render.
Only the alpha clip and alpha to coverage transparency should be allowed
to use theirs shader for the shadow render because these both mode do
binary test which can be made also in the depth map.
But previously the test which added a material in the list of materials
allowed to use their shader was wrong because it only chacked if any
transparency mode were enabled.
To fix this issue KX_BlenderMaterial add the flag RAS_ALPHA_SHADOW to m_rasMode
when alpha clip or alpha to coverage is detected.

Test with 20*20 objects of 2 meshes:
Before: 6.10ms
After: 5.70ms (6.5%)