UPBGE: Avoid sort polygons for shadow render.
Previously the polygons where not sorted only if the drawing mode
was for a wireframe render, finally we could consider that sorting
the polygons is usefull only for a textured render and so a render
not using a override shader. This last condition is now used in
RAS_MeshSlot::RunNode and avoid an useless polygons sort when rendering
shadows.