UPBGE: Simplify bullet occlusion.
The bullet occlusion and culling call was simplified by avoiding passing extra
arguments. The number of planes is statically set to 6, the modelview and projection
are replaced by the matrix from the camera frustum as this matrix is the result
of the same operation made in CullingTest.

In the same time the plane reordering is removed in favor of a reorder in SG_Frustum.