UPBGE: Use VBO and VAO for debug drawing.
Using VBO and VAO allow to render all the lines in one command
and to render all the AABB and frustum using instancing in one
command too.

The usage of VBO request that all the structs for lines, aabbs,
frustum and boxes use array instead of mathfu types to avoid
SSE padding and alignement.

Four shaders are introduced to help debug drawing using instancing.
The first is just for drawing lines using a flat color, the second
is for wire boxes, the third for solid boxes with in and out color
and the fourth for 2d boxes.
All the these geometry type use geometry instancing.

For each type RAS_OpenGLDebugDraw allocate a VAO and multiples VBOs,
they are stored into m_vbos and m_vaos and accessed thanks to enums.
The VAO are initialized in RAS_OpenGlDebugDraw constructor pointing
to the uninitialized VBOs, some VBOs as unit 3d and 2d box and IBOs
are initialized in the same time.

At each render step the VBO containing data about instancing or
geometry are updated from the data list contained into RAS_DebugDraw.
Then the render is procceed binding the VAO and calling the proper
draw call.

2d boxes receive a special treatment. The boxes are rendered using
a vertically inverted ortho matrix because the Y axis is from the
top of the screen.

All these modification ends in a significant speedup for debug drawing
and makes the code easier to move to blender 2.8.