UPBGE: Fix image tile animation.
Previously the Image::lastupdate was to nan after the convertion and
caused the test in GPU_update_image_time to be discarded.
To fix this issue at read time in the function blo_end_image_pointer_map
lastupdate is set to 0.0f.

Bug reported by Spoopy and still in BFBGE.