UPBGE: Remove influence of stereo rasterizer member in matrix compute.
Previously the members m_stereomode and m_curreye were used when computing
the view and projection matrix. But as these settings are not used to change
in game a lot of render loop for texture renderer or shadow were copying the
current stereo mode, change it to avoid its effect in matrix computation and
then restore the stereo mode after. This behaviour is to avoid and to replace
by a send of the stereo mode to the function needing it. These functions are:
GetRenderArea
GetViewMatrix
GetFrustumMatrix
DrawStereoOffScreen

Even if the members are not used in computation, they are left to have a place
to store the global stereo mode to us. Ketsji engine is a candidate but as
these values are only for rendering the rasterizer is a good place.