UPBGE: Fix crash at game exit.
Previously the meshes needed to be freed vefore the scene to ensure that
the bucket manager was still alive, but for doing this we deleted the full
scene slot in the converter before delete the scene.
This as the side effect to delete material and so texture which are needed
in the destruction of the texture renderers. To avoid this dependencie lack
only the meshes are freed before the scene.

Fix issue #539.