UPBGE: Fix conversion of modifier deformer for hidden objects.
Previously the conversion of modifier deformer needed to access to the
game object mesh user to set the derived mesh in its mesh slots.
But in case of the conversion of a hidden object, the mesh user is
invalid. A check is added to avoid acces the mesh slot, we let the
next update of a replicated modifier deformer to set the derived mesh.

Fix issue #540.