UPBGE: Fix deletion of child and parent objects in same time.
Previously if the user ask to delete the parent and child objects
the obht were added to the euthanasy list, then LogicEndFrame we iterated
on these objects and delete them. But when the parent object was deleted
the children were too because the function SG_Node::Destruct() call
recursively a callback calling NewRemoveObject, but in LogicEndFrame
we was still calling a deletion of the child object which create a double
deletion.

To fix this issue the LogicEndFrame only call RemoveObject with the first
element of the euthanasy while it is not empty and in NewRemoveObject the
object is removed from the euthanasy list if it was present.

Fix issue #541.

(cherry picked from commit 695afa08d7bc7aeedf428fc8d9389738d9c7ed63)