UPBGE: Implement PCF jitter shadows;
The jitter PCF is a technic to rendomize the position of the samples
averaged in PCF. To randomize a 64x64 texture with random value between
-1 and 1 is passed to shader. Into the new GLSL function test_shadow_pcf_jitter
we found the rand value at the fragment position in the shadow texture
mutliplied by a big dummy value. Then a matrix is constructed with the
rand values and used to transform the samples offset.

The jitter texture is initialized globally in gpu_texture.cpp, it also
now uses filtering and so the SSAO shader disable it to simulate a texture
without filtering.