UPBGE: Don't render frame when logic step request exit.
Previously when the logic step requested an exit the render
frame just after was proceeded, this wans't a problem in most
of case but in case the logic step removed all the frame
and produce an exit the render of "nothing" will be tried.
Currently this produce a segfault due to the access to the first
frame to get world background info for the first viewport clear.
Also this behaviour made the exit not immediate.

Instead of fixing the render of no scene by discard in Render()
i prefered to not call it to avoid any futur issue. Now we ensure
that the Render() have something to render.

This new behaviour is introduced by checking the exit flag just
before render the frame and do nothing if the flag is not set
to KX_ExitRequest::NO_REQUEST.

Fix issue: #549.