temp work on motion blur
This is a quick and dirt adaptation of eevee's motion blur (camera based
motion blur -> So it doesn't replace old bge per object motion blur with accum
buffer (deprecated))

In eevee iirc it is just for camera's animation so I used eevee's shader
but maybe it is not appropriated

I don't know if the effect is correctly computed. I just quickly did
that... There is some blu but maybe a bit weird.

We could have our own shader/code for this. It is just a test to see if we
can do something interesting with this