UPBGE: Fix python setUseExternalTime call.
Python can't convert a boolean argument from python to
a bool in C through PyArg_ParseTuple. Else a stack smashing
appears (maybe writing a int to a bool and resulting in overflow).
This error is so easy to make and nothing could show it at compilation.
Be aware of this.