A working version of eevee's AO
We still haven't exactly the same render as in the viewport.

But with "normal" settings the results are very similar.

- Maybe the depth texture formats can explain a part of this difference.
- I also noticed that when we increase AO distance too much, the result in
  the viewport is not the same than in bge (maybe distance is clamped
  somewhere in eevee's code?)

I think even if the results are not exactly the same, we can keep it like
that for now.

test file: http://pasteall.org/blend/index.php?id=47381

WARNING: Sometimes AO fails to launch at ge start. I don't know why for
now