UPBGE: Remove SG_[D/Q]List for sensor and fix upper cast.
Previously the SG_DList inherited in SCA_ISensor was used only
in SCA_EventManager. But it's usage create issue when the m_sensors
list in the event manager was empty because SG_DList::iterator do
upper cast anyway even if there's no elements and that the vptr
are invalid. Normal run produce no error but memory check does.

This issue is fixed by using a simple std::vector and checking for
double before adding and removing.