UPBGE: Fix uv and color layer conversion.
Previously the uv and color were set to zero in GetUvRgba based
on the current index and had the side effect to smash the previous
uv layer if a color layer existed and opposite.

To avoid this we follow the fact that all the layers are detect and
use an index smaller than the total count of layers. So all the
layer that will be converted are all the layer that will be needed
and the initialization to zero is obsolete.

Fix issue reported by @youle.