UPBGE: Fix Texture mipmap with ImageRender.
In VideoTexture module, the source can behave in two way:
they return a pointer on a pixel array computed in CPU or
copy a OpenGL texture to the provided bind code.

In the first case the data send to a OpenGL texture work and
detect mipmap and generate them.
In the sceond case the mipmap are just forgotten...

So to fix this issue the calcViewport function, the only function
doing texture to texture copy receive the mipmap settings.
This settings help to generate texture seting for mipmapping
in loadTexture and to call glGenerateMipmap after calling
glCopyTexSubImage2D.
In consideration the caller of calcViewport: getImage, calcImage
have to accept the mipmap argument too.

Fix issue: #528.