UPBGE: Fix general render area dimensions.
Previously multiple issue related to the render area were noticed.
They are the all due to a common issue: the width and height get
in RAS_Rect return the number of pixel between the corner (x1 - x2)
giving the maximum pixel index but not the real number of pixels
which is one pixel more.
This issue wasn't unknown as "+ 1" were spread in all the code for
the functions SetViewport and SetScissor of the rasterizer.

But the python function to get window size didn't used the + 1 and
gave wrong values to the user compared to filter off screen size
which were correct.

To avoid all futur issues the functions GetWidth and GetHeight in
RAS_Rect include the addition of 1 and the + 1 workaround is removed
is the rest of the code.

After testing with a custom shape cursor (a simple dot), it seems that linux
cursor doesn't use the corner of the cursor but the corner without the little
white margin contrary to windows which use the corner with the margin.
The mouse event code was introducing a offset for linux cursor, as this
offset is wrong in windows and the window dimensions are right, the mouse
offset is removed.

Fix issue: #560.