UPBGE: Fix mesh conversion without uv and/or uv layers.
Even without uv or color layers, the meshes are converted with
at least one uv and color layer. The refactor to support MLoop
data conversion removed this behaviour accidentally.

This behaviour is now reset by passing the uv and color layer
count in GetUvRgba to detect if there's no layer and then use
zero vector for uv and a unit vector for color.

In the same time the tangent computation was broken because it
was called in every case even if the mesh doesn't have a set
uv layer. This ended in a assert in blender code.

To fix it, we let the tangent to nullptr when no uv layer are
detected.

Note: RAS_ITexVert::closeTo request the vertex to have the same
format, an assert check is added for.

Thanks to youle for bug report and pre work.
Fix issue: #559.