improve banding effect correction in tonemap filter
We need more tests for that with someone who has the eye for these things.
We could also try eevee tonemapping to see if really it has really a big
performances cost

I think the bands appeared again when we changed the default format for
textures (R11G11B10 instead of RGBA8)

Now we have a problem with "light energy" to fix. Not related to this
branch I think. I think this concerns only volumetrics. I'll test later