fix crash when use shadow is not enabled
Here the problem was that when we don't render shadows,
RAS_ManagerNodeData::m_overrideShader (pointer) had no default value so
when we executed this code:

m_nodeData.m_shader = overrideShader ? overrideShader : m_shader;

in RAS_MaterialBucket::GenerateTree, override shader had 0xcdcdcdcd value
which is not considered as nullptr with VisualStudio so
m_nodeData.m_shader was equal to overrideShader and overrideShader was
empty (because we were not running a shadow pass (we were running the
RAS_TEXTURED pass)). This caused a crash later in the code.

So all pointers in RAS_RenderNode.h must have a default value set to
nullptr to avoid this kindof issues