UPBGE: Fix object with timer deleted before timer end.
Previously the object in m_tempObjectList were removed only
when the timer ended, but if the user ask to delete the object
before the timer does, then the object still remains in
m_tempObjectList and at the next logic frame we will intent
to read the timer property from all the object in the list
and we will get invalid pointers.

To solve this issue the object is removed from the temporary
list only in NewRemoveObject.

Fix issue: #565.