begin to work on eevee launching process refactor
I tried another approach to launch bge from eevee. The principle is to
render eevee scene before launching without freeing data so we can access
it later.

The problem is that I don't know how to do that for all scenes.

I also removed m_viewport in RAS_Rasterizer but I don't know if it is
good. I maybe broke the correct size of the viewport area when we are
looking threw camera. So maybe we'd need to revert the changes I did here.